requirements.txt
gunicorn==20.1.0
dash==2.6.1
pandas==1.1.5
dash-bootstrap-components==1.2.1
pyxnat==1.4
dash-uploader==0.6.0
packaging==21.0 # necessary for dash-uploader 0.6.0
dash-slicer==0.3.0
nilearn==0.9.2
flask-login==0.6.
python-dotenv==0.20.0
requests==2.28.1
